Fixing Crashes, Hardcore Mode, and Fixing Encoding Errors
1. EVERY Event is now in ANSI. Before, several were in UTF-EIGHT and
others were in UTF-EIGHT WITHOUT BOM. It's a shame that the Cyrillic
entries took a hit, but this has been known to cause crashes before and
was neccesary.

2. Per Had's request, the flavorful modifiers given to a nation that
takes the Exodus are now tied to whether or not Hardcore Mode is enabled
or disabled.

3. Pirates no longer steal Anamolies by surveying. Why does a pirate
care about cataloging the life, geography, and biology of a planet?

4. The earth is an Early Space Age Primitive state right now. We want to
colonize Mars, which has zero Habitability. So too should Primitives be
allowed to colonize crappy planets.

5. Xenophiles, longing for all Pops to be free, will now try and buy
Pirate Slaves if they can