Fixed Robots Becoming Organic Sub-Species Bug
Continuing from the last commitment, I have found that HFE, when using
the commands to turn Machine Unit Robots in to Mechanical Robots, and
vice-versa, it doesn't check if the primary species is organic.

I also fixed a bug involving three IF commands that did nothing if their
LIMITS were fulfilled.