Encoding Error
The "HFE_ANOMALY.TXT" file was encoded in UTF-Eight, thereby causing an
immediate crash to the desktop.  Paradox uses ANSI, and now this file
will too.

I also removed some false positives in the Error Log concerning the
Pirate Civics.

The AI can now get the Blight.

If Hardcore Mode is disabled, there remains a chance that the AI will
get the Blight!