Delete message_types.txt
Had, the reason I cut the Pirate Notification Message Types is because you (not me) changed the files so that it spawns a dynamic notification. Said dynamic notification need not replace a Vanilla rendition file. That whole file was meant to be deleted.

(I also changed back the alert to "hfe_advisor_notification_pirate_raid".)